chore: copy location warnings to settings page
#1265
