+++ vrc-get-gui/app/(setup)/setup/backups/page.tsx
@@ -2,6 +2,7 @@
 
 import {
 	BackupFormatSelect,
+	BackupPathWarnings,
 	FilePathRow,
 } from "@/components/common-setting-parts";
 import { CardDescription } from "@/components/ui/card";
@@ -12,11 +13,7 @@ import {
 import { useGlobalInfo } from "@/lib/global-info";
 import { tc } from "@/lib/i18n";
 import { toastThrownError } from "@/lib/toast";
-import {
-	type BodyProps,
-	SetupPageBase,
-	WarningMessage,
-} from "../setup-page-base";
+import { type BodyProps, SetupPageBase } from "../setup-page-base";
 
 export default function Page() {
 	const isMac = useGlobalInfo().osType === "Darwin";
@@ -46,14 +43,6 @@ function Body({ environment, refetch }: BodyProps) {
 		}
 	};
 
-	const globalInfo = useGlobalInfo();
-	const isWindows = globalInfo.osType === "WindowsNT";
-	const inLocalAppData = !!(
-		isWindows &&
-		globalInfo.localAppData &&
-		projectBackupPath.includes(globalInfo.localAppData)
-	);
-
 	return (
 		<>
 			<h3>{tc("setup:backups:location")}</h3>
@@ -67,11 +56,7 @@ function Body({ environment, refetch }: BodyProps) {
 				refetch={refetch}
 				successMessage={tc("settings:toast:backup path updated")}
 			/>
-			{inLocalAppData && (
-				<WarningMessage>
-					{tc("setup:backups:warning:in-local-app-data")}
-				</WarningMessage>
-			)}
+			<BackupPathWarnings backupPath={projectBackupPath} />
 			<div className={"pb-3"} />
 			<h3>{tc("setup:backups:archive")}</h3>
 			<CardDescription className={"whitespace-normal"}>
