+++ vrc-get-gui/app/(setup)/setup/project-path/page.tsx
@@ -1,15 +1,13 @@
 "use client";
 
-import { FilePathRow } from "@/components/common-setting-parts";
+import {
+	FilePathRow,
+	ProjectPathWarnings,
+} from "@/components/common-setting-parts";
 import { CardDescription } from "@/components/ui/card";
 import { environmentPickProjectDefaultPath } from "@/lib/bindings";
-import { useGlobalInfo } from "@/lib/global-info";
 import { tc } from "@/lib/i18n";
-import {
-	type BodyProps,
-	SetupPageBase,
-	WarningMessage,
-} from "../setup-page-base";
+import { type BodyProps, SetupPageBase } from "../setup-page-base";
 
 export default function Page() {
 	return (
@@ -24,19 +22,6 @@ export default function Page() {
 }
 
 function Body({ environment, refetch }: BodyProps) {
-	const projectPath = environment.default_project_path;
-
-	const hasWhitespace = projectPath.includes(" ");
-	const globalInfo = useGlobalInfo();
-	const isWindows = globalInfo.osType === "WindowsNT";
-	// biome-ignore lint/suspicious/noControlCharactersInRegex: allow control characters
-	const hasNonAscii = isWindows && projectPath.match(/[^\x00-\x7F]/);
-	const inLocalAppData = !!(
-		isWindows &&
-		globalInfo.localAppData &&
-		projectPath.includes(globalInfo.localAppData)
-	);
-
 	return (
 		<>
 			<CardDescription className={"whitespace-normal"}>
@@ -44,28 +29,12 @@ function Body({ environment, refetch }: BodyProps) {
 			</CardDescription>
 			<FilePathRow
 				withoutSelect
-				path={projectPath}
+				path={environment.default_project_path}
 				pick={environmentPickProjectDefaultPath}
 				refetch={refetch}
 				successMessage={tc("settings:toast:default project path updated")}
 			/>
-			<div className="flex flex-col gap-1">
-				{hasWhitespace && (
-					<WarningMessage>
-						{tc("setup:project-path:warning:whitespace")}
-					</WarningMessage>
-				)}
-				{hasNonAscii && (
-					<WarningMessage>
-						{tc("setup:project-path:warning:non-ascii")}
-					</WarningMessage>
-				)}
-				{inLocalAppData && (
-					<WarningMessage>
-						{tc("setup:project-path:warning:in-local-app-data")}
-					</WarningMessage>
-				)}
-			</div>
+			<ProjectPathWarnings projectPath={environment.default_project_path} />
 		</>
 	);
 }
