+++ vrc-get-gui/locales/en.json5
@@ -249,6 +249,11 @@
     "general:button:select": "Select",
     "settings:unity installations": "Unity Installations",
     "settings:button:add unity": "Add Unity",
+    // Reset your PC is a feature in Windows, so please check your translation is consistent with Windows terminology.
+    // It's at Setting app > System > Recovery > Reset this PC
+    "settings:warning:in-local-app-data": "The location is in LocalAppData folder, which will be deleted with \"Reset this PC\" with \"Keep my files\".<br>It's recommended to save your data in a different location.",
+    "settings:warning:whitespace": "The location contains whitespace. Whitespace in may cause problems with Unity and other tools. Please consider use another location.",
+    "settings:warning:non-ascii": "The location contains non-ASCII characters. Non-ASCII characters in the location may cause problems with Unity. Please consider use another location.",
     "settings:default project path": "Default Project Path",
     "settings:default project path description": "The default project path is the directory where new projects are created.",
     "settings:backup:path": "Backup Path",
@@ -327,15 +332,9 @@
     "setup:unity-hub:not found": "Unity Hub was not found.",
     "setup:project-path:heading": "Project Save Location",
     "setup:project-path:description": "When you create project with ALCOM, the project will be saved in this location.",
-    "setup:project-path:warning:whitespace": "The location contains whitespace. Whitespace in may cause problems with Unity and other tools. Please consider use another location.",
-    "setup:project-path:warning:non-ascii": "The location contains non-ASCII characters. Non-ASCII characters in the location may cause problems with Unity. Please consider use another location.",
-    // Reset your PC is a feature in Windows, so please check your translation is consistent with Windows terminology.
-    // It's at Setting app > System > Recovery > Reset this PC
-    "setup:project-path:warning:in-local-app-data": "The location is in LocalAppData folder, which will be deleted with \"Reset this PC\" with \"Keep my files\".<br>It's recommended to save your projects in a different location.",
     "setup:backups:heading": "Project Backup Settings",
     "setup:backups:location": "Backup Location",
     "setup:backups:location description": "When you backup your project, the backup archive file will be saved in this path.",
-    "setup:backups:warning:in-local-app-data": "The location is in LocalAppData folder, which will be deleted with \"Reset this PC\" with \"Keep my files\".<br>It's recommended to save your backups in a different location.",
     "setup:backups:archive": "Archive Format",
     "setup:backups:archive description": "Backups are saved as an archive file. <br/>You can select the format of the backup archive. <br/>Default setting is the best format ALCOM thinks at the moment and current default behavior is represented in braces.<br/>Compressing the archive will make backup took longer but will save disk space.",
     "setup:system-setting:heading": "System Configurations",
